requirements.txt
PySide6>=6.5.0
blake3>=0.4.1
PyYAML